refactor: MenuService 에서 productRepository 에 대한 의존성 제거
MenuProductTranslator 인터페이스를 통해 추상화
