+++ src/main/java/kitchenpos/menus/tobe/infra/MenuProductTranslator.java
@@ -1,0 +1,10 @@
+package kitchenpos.menus.tobe.infra;
+
+import kitchenpos.menus.tobe.domain.MenuProducts;
+import kitchenpos.menus.tobe.dto.MenuProductRequest;
+
+import java.util.List;
+
+public interface MenuProductTranslator {
+    MenuProducts translateMenuProducts(List<MenuProductRequest> menuProductRequests);
+}
