+++ src/main/java/kitchenpos/products/tobe/infra/ApiProductTranslator.java
@@ -1,6 +1,5 @@
-package kitchenpos.common.infra;
+package kitchenpos.products.tobe.infra;
 
-import kitchenpos.products.tobe.domain.ProductTranslator;
 import kitchenpos.products.tobe.dto.UpdateMenuStatusRequest;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
