+++ src/main/java/kitchenpos/products/tobe/infra/ProductTranslator.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.infra;
 
 import java.util.UUID;
