+++ src/main/java/kitchenpos/products/tobe/infra/ServiceProductTranslator.java
@@ -1,7 +1,6 @@
-package kitchenpos.common.infra;
+package kitchenpos.products.tobe.infra;
 
 import kitchenpos.menus.tobe.application.MenuService;
-import kitchenpos.products.tobe.domain.ProductTranslator;
 
 import java.util.UUID;
